569: Renamed SwapDetail to OrderDetail
