+++ src/pages/OrderDetail/OrderDetail.tsx
@@ -2,8 +2,8 @@ import React, { FC } from "react";
 
 import Page from "../../components/Page/Page";
 
-const SwapDetail: FC = () => {
+const OrderDetail: FC = () => {
   return <Page>swap detail page (swap-id/:chain/:nonce)</Page>;
 };
 
-export default SwapDetail;
+export default OrderDetail;
